+++ assets/src/admin/components/footer.js
@@ -1,6 +1,6 @@
 import { __ } from '@wordpress/i18n'
 import { Button } from '@wordpress/components'
-import useSettings from '../use-settings'
+import useSettings from '../hooks/use-settings'
 
 /**
  * Sticky footer with the save button.
