+++ assets/src/admin/tabs/general/index.js
@@ -1,0 +1,16 @@
+import LoadingBehaviour from './loading-behaviour'
+import LoadButton from './load-button'
+import Cache from './cache'
+
+/**
+ * General settings tab — loading behaviour + cache management.
+ */
+const General = () => (
+	<>
+		<LoadingBehaviour />
+		<LoadButton />
+		<Cache />
+	</>
+)
+
+export default General
