+++ assets/src/settings.js
@@ -1,10 +1,10 @@
 /**
  * Admin settings app entry point.
  */
-import './styles/settings.scss'
+import './admin/styles/settings.scss'
 import domReady from '@wordpress/dom-ready'
 import { createRoot } from '@wordpress/element'
-import App from './admin/app'
+import App from './admin'
 
 domReady(() => {
 	const el = document.getElementById('lazy-load-for-comments-settings')
